refactor: move ordering-related code to `order_by_util.rs`
